fix(core): fix addDecoToClass params;remove Nested
