+++ .changeset/late-squids-decide.md
@@ -1,0 +1,8 @@
+---
+"phecda-core": patch
+---
+
+refactor Err logic;
+remove partial on transformClass;
+remove Nested
+change addDecoToClass params
