+++ packages/core/src/decorators/index.ts
@@ -1,0 +1,2 @@
+export * from './core'
+export * from './function'
