+++ packages/core/src/helper.ts
@@ -1,5 +1,5 @@
 /* eslint-disable new-cap */
-import { getExposeKey, getHandler, getModuleState, getState } from './core'
+import { SHARE_KEY, getExposeKey, getHandler, getModuleState, getState } from './core'
 import type { AbConstruct, ClassValue, Construct, Phecda } from './types'
 
 export function getTag<M extends Construct | AbConstruct>(moduleOrInstance: M | InstanceType<M>): PropertyKey {
@@ -34,15 +34,13 @@ export function plainToClass<M extends Construct, Data extends Record<PropertyKe
   return instance
 }
 
-export async function transformClass<M extends Construct>(instance: InstanceType<M>, force = false, partial = false) {
+export async function transformClass<M extends Construct>(instance: InstanceType<M>, force = false) {
   const err: string[] = []
 
   const stateVars = getModuleState(instance) as PropertyKey[]
   for (const item of stateVars) {
-    if (partial && instance[item] === undefined)
-      continue
-
     const handlers = getHandler(instance, item)
+
     if (handlers) {
       for (const handler of handlers) {
         const pipe = handler.pipe
@@ -90,8 +88,8 @@ export function snapShot<T extends Construct>(data: InstanceType<T>) {
 /**
  * add decorator to a class by function
  */
-export function addDecoToClass<M extends Construct | AbConstruct>(c: M, key: keyof InstanceType<M> | string, handler: PropertyDecorator | ClassDecorator, type: 'property' | 'class' = 'class') {
-  handler(type === 'class' ? c.prototype : c, key as any)
+export function addDecoToClass<M extends Construct | AbConstruct>(c: M, key: keyof InstanceType<M> | PropertyKey, handler: PropertyDecorator | ClassDecorator) {
+  handler(key === SHARE_KEY ? c : c.prototype, key as any)
 }
 
 export function Pipeline(...decos: ((...args: any) => void)[]) {
@@ -99,4 +97,8 @@ export function Pipeline(...decos: ((...args: any) => void)[]) {
     for (const d of decos)
       d(...args)
   }
+}
+
+export function isAsyncFunc(fn: Function) {
+  return (fn as any)[Symbol.toStringTag] === 'AsyncFunction'
 }
