+++ packages/core/src/index.ts
@@ -2,5 +2,5 @@ export * from './decorators'
 export * from './types'
 export * from './helper'
 export * from './core'
-export * from './custom/decorator'
 // export * from './preset'
+export * from './di'
