+++ packages/server/src/server/fastify/bind.ts
@@ -101,8 +101,6 @@ export function bindApp(app: FastifyInstance, { moduleMap, meta }: Awaited<Retur
               const {
                 paramsType,
 
-                handlers,
-
                 data: {
                   params,
                   guards,
@@ -141,7 +139,6 @@ export function bindApp(app: FastifyInstance, { moduleMap, meta }: Awaited<Retur
                 resolve(await context.usePostInterceptor(funcData))
               }
               catch (e: any) {
-                handlers.forEach(handler => handler.error?.(e, context.data))
                 resolve(await context.useFilter(e, filter))
               }
             })
@@ -167,7 +164,6 @@ export function bindApp(app: FastifyInstance, { moduleMap, meta }: Awaited<Retur
 
       const {
         paramsType,
-        handlers,
         data: {
           interceptors,
           guards,
@@ -223,7 +219,6 @@ export function bindApp(app: FastifyInstance, { moduleMap, meta }: Awaited<Retur
             return ret
           }
           catch (e: any) {
-            handlers.forEach(handler => handler.error?.(e, context.data))
             const err = await context.useFilter(e, filter)
 
             if (res.sent)
