+++ packages/server/src/server/koa/bind.ts
@@ -87,7 +87,6 @@ export function bindApp(app: Router, { moduleMap, meta }: Awaited<ReturnType<typ
             const [name, method] = tag.split('-')
             const {
               paramsType,
-              handlers,
               data: {
                 params,
                 guards, interceptors,
@@ -122,7 +121,6 @@ export function bindApp(app: Router, { moduleMap, meta }: Awaited<ReturnType<typ
               resolve(await context.usePostInterceptor(funcData))
             }
             catch (e: any) {
-              handlers.forEach(handler => handler.error?.(e, context.data))
               resolve(await context.useFilter(e, filter))
             }
           })
@@ -144,7 +142,6 @@ export function bindApp(app: Router, { moduleMap, meta }: Awaited<ReturnType<typ
 
       const {
         paramsType,
-        handlers,
         data: {
           interceptors,
           guards,
@@ -195,7 +192,6 @@ export function bindApp(app: Router, { moduleMap, meta }: Awaited<ReturnType<typ
           ctx.body = ret
         }
         catch (e: any) {
-          handlers.forEach(handler => handler.error?.(e, context.data))
           const err = await context.useFilter(e, filter)
 
           if (ctx.res.writableEnded)
