+++ packages/vue/__test__/vue.spec.ts
@@ -22,35 +22,38 @@ describe('work for vue', () => {
     expect(name.value).toBe('phecda')
   })
   it('error handler', async () => {
-    const fn = vi.fn((info: string) => `info:${info}`)
+    const fn = vi.fn((err: Error) => `${err.message}`)
     class ErrorMaker {
       name: string
       testA(param: any) {
         throw new Error(param)
       }
 
-      @Err(fn)
-      testB(param: any) {
-        throw new Error(param)
+      @Err(fn, true)
+      invoke() {
+        throw new Error('invoke error')
       }
 
-      @Err(fn)
-      async testC(param: any) {
-        return Promise.reject(param)
+      @Err(fn, false)
+      async throw() {
+        // eslint-disable-next-line prefer-promise-reject-errors
+        return Promise.reject('throw error')
       }
     }
     // work with useR
     const ret = useR(ErrorMaker)
-    expect(() => ret.testA('A')).toThrowError('A')
-    expect(ret.testB('B')).toBe('info:Error: B')
-    expect(await ret.testC('C')).toBe('info:C')
+    ret.invoke()
+    expect(fn).toHaveLastReturnedWith('invoke error')
+    await expect(ret.throw()).rejects.toThrow('throw error')
     expect(fn).toHaveBeenCalledTimes(2)
 
     // work with useV
-    const { testA, testB, testC } = useV(ErrorMaker)
-    expect(() => testA('A')).toThrowError('A')
-    expect(testB('B')).toBe('info:Error: B')
-    expect(await testC('C')).toBe('info:C')
+    const { invoke, throw: ThrowError } = useV(ErrorMaker)
+
+    invoke()
+    expect(fn).toHaveLastReturnedWith('invoke error')
+    await expect(ThrowError()).rejects.toThrow('throw error')
+
     expect(fn).toHaveBeenCalledTimes(4)
   })
