+++ packages/vue/src/composable.ts
@@ -3,13 +3,12 @@ import type { UnwrapNestedRefs } from 'vue'
 import { onBeforeUnmount, reactive, shallowReactive, toRaw, toRef } from 'vue'
 
 import type { Construct, Events } from 'phecda-web'
-import { emitter, getActiveInstance, getHandler, getTag, invokeHandler, wrapError } from 'phecda-web'
+import { emitter, getActiveInstance, getTag, invokeHandler } from 'phecda-web'
 import type { ReplaceInstanceValues } from './types'
 import type { DeepPartial } from './utils'
 import { createSharedReactive, mergeReactiveObjects } from './utils'
 
 const REF_SYMBOL = Symbol('ref')
-const REACTIVE_SYMBOL = Symbol('reactive')
 
 export function useO<T extends Construct>(module: T): UnwrapNestedRefs<InstanceType<T>> {
   const { state, origin } = getActiveInstance()
@@ -50,39 +49,7 @@ export function usePatch<T extends Construct>(module: T, Data: DeepPartial<Insta
 }
 
 export function useR<T extends Construct>(module: T): UnwrapNestedRefs<InstanceType<T>> {
-  const { cache: cacheMap } = getActiveInstance()
-  const instance = useO(module)
-
-  const cache = cacheMap.get(instance) || {}
-
-  if (cache[REACTIVE_SYMBOL])
-    return cache[REACTIVE_SYMBOL]
-
-  const proxy = new Proxy(instance, {
-    get(target: any, key) {
-      if (typeof target[key] === 'function') {
-        if (cache[key])
-          return cache[key]
-        const errorHandler = getHandler(target, key).find((item: any) => item.error)?.error
-        if (!errorHandler)
-          return target[key].bind(target)
-        const wrapper = wrapError(target, key, errorHandler)
-        cache[key] = wrapper
-        return wrapper
-      }
-
-      return target[key]
-    },
-    set(target: any, key, v) {
-      target[key] = v
-      return true
-    },
-  })
-
-  cache[REACTIVE_SYMBOL] = proxy
-  if (!cacheMap.has(instance))
-    cacheMap.set(instance, cache)
-  return proxy
+  return useO(module)
 }
 
 export function useV<T extends Construct>(module: T): ReplaceInstanceValues<InstanceType<T>> {
@@ -94,17 +61,10 @@ export function useV<T extends Construct>(module: T): ReplaceInstanceValues<Inst
     return cache[REF_SYMBOL]
   const proxy = new Proxy(instance, {
     get(target: any, key) {
-      if (typeof target[key] === 'function') {
-        if (cache[key])
-          return cache[key]
-        const errorHandler = getHandler(target, key).find((item: any) => item.error)?.error
-        if (!errorHandler)
-          return target[key].bind(target)
-        const wrapper = wrapError(target, key, errorHandler)
-        cache[key] = wrapper
-        return wrapper
-      }
-      if (target[key]?.__v_skip)
+      if (typeof target[key] === 'function')
+        return target[key].bind(this)
+
+      if (target[key]?.__v_skip)// markRaw
         return target[key]
 
       const cacheRef = cache[key]
