+++ packages/web/src/index.ts
@@ -2,4 +2,3 @@ export * from 'phecda-core'
 export * from './types'
 export * from './plugin'
 export * from './core'
-export * from './utils'
